Update Readme and Reorganise Menu
