+++ src/app/app.component.html
@@ -96,10 +96,10 @@
             Misc
           </a>
           <div class="dropdown-menu" aria-labelledby="navbarRules">
-            <a class="dropdown-item" routerLink="/misc-org-time-update">Set Org Time</a>
-            <a class="dropdown-item" routerLink="/misc-manage-pat">Manage PAT</a>
+            <a class="dropdown-item" routerLink="/entitlement-owners">Manage Entitlements</a>
             <a class="dropdown-item" routerLink="/identity-admin-management">Manage Admins</a>
-            <a class="dropdown-item" routerLink="/entitlement-owners">Manage Entitlement Owners</a>
+            <a class="dropdown-item" routerLink="/misc-manage-pat">Manage PAT</a>
+            <a class="dropdown-item" routerLink="/misc-org-time-update">Set Org Time</a>
             <a class="dropdown-item" routerLink="/misc-org-stats">Org Statistics</a>
           </div>
         </li>
