+++ src/app/app.module.ts
@@ -33,7 +33,7 @@ import { IdentityLCSComponent } from './identity-lcs-management/identity-lcs-man
 import { MultipleAccountsComponent } from './multiple-accounts-report/multiple-accounts-report.component';
 import { WorkItemsStatusComponent } from './work-items-status/work-items-status.component';
 import { WorkItemsForwardComponent } from './work-items-forward/work-items-forward.component';
-import { EntitlmentOwnersComponent } from './misc-entitlement-owners/misc-entitlement-owners.component';
+import { EntitlmentOwnersComponent } from './misc-manage-entitlements/misc-manage-entitlements.component';
 import { ResetSourceComponent } from './source-reset/source-reset.component';
 import { ChangeSourceOwnerComponent } from './source-owner-update/source-owner-update.component';
 import { SourceCreateProfileComponent } from './source-create-profile/source-create-profile.component';
