+++ src/app/misc-manage-entitlements/misc-manage-entitlements.component.html
@@ -2,7 +2,7 @@
   <ol class="breadcrumb m-0">
     <li class="breadcrumb-item"><a class="text-decoration-none" routerLink="/identity-info">Home</a></li>
     <li class="breadcrumb-item">Misc</li>
-    <li class="breadcrumb-item active" aria-current="page">Manage Entitlement Owners</li>
+    <li class="breadcrumb-item active" aria-current="page">Manage Entitlements</li>
   </ol>
 </nav>
 <div>
