+++ src/app/misc-manage-entitlements/misc-manage-entitlements.component.spec.ts
@@ -1,6 +1,6 @@
 import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
-import { EntitlmentOwnersComponent } from './misc-entitlement-owners.component';
+import { EntitlmentOwnersComponent } from './misc-manage-entitlements.component';
 import { HttpClientModule } from '@angular/common/http';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { FormsModule } from '@angular/forms';
